+++ airflow/contrib/hooks/aws_dynamodb_hook.py
@@ -16,59 +16,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+"""This module is deprecated. Please use `airflow.providers.amazon.aws.hooks.aws_dynamodb_hook`."""
 
+import warnings
 
-"""
-This module contains the AWS DynamoDB hook
-"""
-from airflow.exceptions import AirflowException
-from airflow.providers.amazon.aws.hooks.aws_hook import AwsHook
+# pylint: disable=unused-import
+from airflow.providers.amazon.aws.hooks.aws_dynamodb_hook import AwsDynamoDBHook  # noqa
 
-
-class AwsDynamoDBHook(AwsHook):
-    """
-    Interact with AWS DynamoDB.
-
-    :param table_keys: partition key and sort key
-    :type table_keys: list
-    :param table_name: target DynamoDB table
-    :type table_name: str
-    :param region_name: aws region name (example: us-east-1)
-    :type region_name: str
-    """
-
-    def __init__(self,
-                 table_keys=None,
-                 table_name=None,
-                 region_name=None,
-                 *args, **kwargs):
-        self.table_keys = table_keys
-        self.table_name = table_name
-        self.region_name = region_name
-        self.conn = None
-        super().__init__(*args, **kwargs)
-
-    def get_conn(self):
-        self.conn = self.get_resource_type('dynamodb', self.region_name)
-        return self.conn
-
-    def write_batch_data(self, items):
-        """
-        Write batch items to DynamoDB table with provisioned throughout capacity.
-        """
-
-        dynamodb_conn = self.get_conn()
-
-        try:
-            table = dynamodb_conn.Table(self.table_name)
-
-            with table.batch_writer(overwrite_by_pkeys=self.table_keys) as batch:
-                for item in items:
-                    batch.put_item(Item=item)
-            return True
-        except Exception as general_error:
-            raise AirflowException(
-                'Failed to insert items in dynamodb, error: {error}'.format(
-                    error=str(general_error)
-                )
-            )
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.hooks.aws_dynamodb_hook`.",
+    DeprecationWarning, stacklevel=2
+)
