+++ airflow/contrib/operators/dynamodb_to_s3.py
@@ -31,8 +31,8 @@ from uuid import uuid4
 
 from boto.compat import json  # type: ignore
 
-from airflow.contrib.hooks.aws_dynamodb_hook import AwsDynamoDBHook
 from airflow.models import BaseOperator
+from airflow.providers.amazon.aws.hooks.aws_dynamodb_hook import AwsDynamoDBHook
 from airflow.providers.amazon.aws.hooks.s3 import S3Hook
