+++ tests/contrib/operators/test_hive_to_dynamodb_operator.py
@@ -27,7 +27,7 @@ import pandas as pd
 
 import airflow.contrib.operators.hive_to_dynamodb
 from airflow import DAG
-from airflow.contrib.hooks.aws_dynamodb_hook import AwsDynamoDBHook
+from airflow.providers.amazon.aws.hooks.aws_dynamodb_hook import AwsDynamoDBHook
 
 DEFAULT_DATE = datetime.datetime(2015, 1, 1)
 DEFAULT_DATE_ISO = DEFAULT_DATE.isoformat()
