+++ tests/test_core_to_contrib.py
@@ -410,6 +410,10 @@ HOOK = [
         'airflow.providers.amazon.aws.hooks.aws_hook.AwsHook',
         'airflow.contrib.hooks.aws_hook.AwsHook',
     ),
+    (
+        'airflow.providers.amazon.aws.hooks.aws_dynamodb_hook.AwsDynamoDBHook',
+        'airflow.contrib.hooks.aws_dynamodb_hook.AwsDynamoDBHook',
+    ),
 ]
 
 OPERATOR = [
